# 8298639: Perform I/O operations in bulk for RandomAccessFile