--- a/test/jdk/java/io/File/Basic.java
+++ b/test/jdk/java/io/File/Basic.java
@@ -29,10 +29,8 @@
  */
 
 import java.io.FileOutputStream;
-import java.io.IOException;
 import java.io.File;
 import java.io.PrintStream;
-import java.io.RandomAccessFile;
 
 
 public class Basic {