This PR suggests improving performance for read and write operations for the longer primitives in `RandomAccessFile`.

These improvements would also propagate to other JDK classes relying on these improved `RandomAccessFile` operations.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8298639](https://bugs.openjdk.org/browse/JDK-8298639): Perform I/O operations in bulk for RandomAccessFile


### Reviewers
 * [Alan Bateman](https://openjdk.org/census#alanb) (@AlanBateman - **Reviewer**)
 * [Brian Burkhalter](https://openjdk.org/census#bpb) (@bplb - **Reviewer**)


### Contributors
 * Sergey Tsypanov `<stsypanov@openjdk.org>`

### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk pull/11644/head:pull/11644` \
`$ git checkout pull/11644`

Update a local copy of the PR: \
`$ git checkout pull/11644` \
`$ git pull https://git.openjdk.org/jdk pull/11644/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 11644`

View PR using the GUI difftool: \
`$ git pr show -t 11644`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/11644.diff">https://git.openjdk.org/jdk/pull/11644.diff</a>

</details>