@AlanBateman 2022-12-13
@stsypanov is working on the same thing in JDK-8292937 / PR10031.