@stsypanov 2022-12-13
As far as I understood https://github.com/openjdk/jdk/pull/10031 was closed because there we doubts about whether it's safe to use shared array