@AlanBateman 2022-12-13
> As far as I understand #10031 was closed because there we doubts about whether it's safe to use shared array within multiple methods. Here we use the same approach.

I think that PR was closed as it went inactive but I think where you got to is very similar to what Per's change is now.

RandomAccessFile doesn't specify it but it can't be used from several threads without external synchronization. This is because it maintains a file position that is updated when reading and writing, or it can be changed when seeking to positions in the file. So on the surface, the byte[] should be okay.