@stsypanov 2022-12-14
> I think that PR was closed as it went inactive

Yep, I think there I did everything I could, hope this one will be merged as the changes are helpful